fix: move producer and consumer in their own files
