+++ cluster/replication/shard_replication_engine.go
@@ -19,16 +19,33 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/cenkalti/backoff/v4"
 	"github.com/sirupsen/logrus"
-	"github.com/weaviate/weaviate/cluster/proto/api"
-	"github.com/weaviate/weaviate/cluster/replication/types"
 	enterrors "github.com/weaviate/weaviate/entities/errors"
 )
 
 const (
 	replicationEngineLogAction = "replication_engine"
 )
+
+// TimeProvider abstracts time operations to enable testing without time dependencies.
+type TimeProvider interface {
+	Now() time.Time
+}
+
+// RealTimeProvider implements the TimeProvider interface using the standard time package
+type RealTimeProvider struct{}
+
+// Now returns the current time
+func (p RealTimeProvider) Now() time.Time {
+	return time.Now()
+}
+
+// Timer defines an interface for scheduling tasks with a delay.
+type Timer interface {
+	// AfterFunc waits for the specified duration to elapse and then calls the provided function
+	// in its own goroutine.
+	AfterFunc(duration time.Duration, fn func()) *time.Timer
+}
 
 // ShardReplicationEngine coordinates the replication of shard data between nodes in a distributed system.
 //
@@ -280,381 +297,4 @@ func (e *ShardReplicationEngine) OpChannelLen() int {
 // the engine associated with the node.
 func (e *ShardReplicationEngine) String() string {
 	return fmt.Sprintf("replication engine on node '%s'", e.nodeId)
-}
-
-// OpProducer is an interface for producing replication operations.
-type OpProducer interface {
-	// Produce starts producing replication operations and sends them to the provided channel.
-	// A buffered channel is typically used for backpressure, but an unbounded channel may cause
-	// memory growth if the consumer falls behind. Errors during production should be returned.
-	Produce(ctx context.Context, out chan<- ShardReplicationOp) error
-}
-
-// OpConsumer is an interface for consuming replication operations.
-type OpConsumer interface {
-	// Consume starts consuming operations from the provided channel.
-	// The consumer processes operations, and a buffered channel is typically used to apply backpressure.
-	// The consumer should return an error if it fails to process any operation.
-	Consume(ctx context.Context, in <-chan ShardReplicationOp) error
-}
-
-// FSMOpProducer is an implementation of the OpProducer interface that reads replication
-// operations from a ShardReplicationFSM, which tracks the state of replication operations.
-type FSMOpProducer struct {
-	logger          *logrus.Entry
-	fsm             *ShardReplicationFSM
-	pollingInterval time.Duration
-	nodeId          string
-}
-
-// String returns a string representation of the FSMOpProducer,
-// including the node ID that uniquely identifies the producer.
-//
-// The assumption is that each node runs one and only one replication engine,
-// which means there is one producer per node.
-func (p *FSMOpProducer) String() string {
-	return fmt.Sprintf("replication engine FSM producer on node '%s'", p.nodeId)
-}
-
-// NewFSMOpProducer creates a new FSMOpProducer instance, which periodically polls the
-// ShardReplicationFSM for operations assigned to the given node and pushes them to
-// a channel for consumption by the replication engine.The polling interval controls
-// how often the FSM is queried for replication operations.
-//
-// Additional configuration can be applied using optional FSMProducerOption functions.
-func NewFSMOpProducer(logger *logrus.Entry, fsm *ShardReplicationFSM, pollingInterval time.Duration, nodeId string) *FSMOpProducer {
-	return &FSMOpProducer{
-		logger:          logger,
-		fsm:             fsm,
-		pollingInterval: pollingInterval,
-		nodeId:          nodeId,
-	}
-}
-
-// Produce implements the OpProducer interface and starts producing operations for the given node.
-func (p *FSMOpProducer) Produce(ctx context.Context, out chan<- ShardReplicationOp) error {
-	p.logger.WithFields(logrus.Fields{"node": p.nodeId, "pollingInterval": p.pollingInterval}).Info("starting replication engine FSM producer")
-
-	ticker := time.NewTicker(p.pollingInterval)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			p.logger.WithFields(logrus.Fields{
-				"producer": p,
-			}).Info("replication engine producer cancel request, stopping FSM producer")
-			return ctx.Err()
-		case <-ticker.C:
-			// Here we get ALL operations for a certain node which the replication engine is responsible for running.
-			// We write all of them to a channel for the OpConsumer to consume them. Replication operations already
-			// started will be detected by checking the existence of the shard on the target node before the right
-			// before starting the replication operation.
-			ops := p.allOpsForNode(p.nodeId)
-			if len(ops) > 0 {
-				p.logger.WithField("opCount", len(ops)).Debug("preparing op replication")
-
-				for _, op := range ops {
-					select {
-					case <-ctx.Done():
-						return ctx.Err()
-					case out <- op: // Write operation to channel
-					}
-				}
-			}
-		}
-	}
-}
-
-// allOpsForNode filters and returns replication operations assigned to the specified node.
-//
-// This method implements the core of the pull-based replication mechanism:
-//
-// 1. Pull Model: Each node is responsible for pulling data TO itself FROM other nodes
-//
-// 2. Node Responsibility:
-//   - Target node: Handles all replication operations which are in REGISTERED or HYDRATING
-//   - Source node: Only handles DEHYDRATING operations as that state needs data to be deleted
-//
-// 3. Operation States:
-//   - REGISTERED: Initial state, operation waiting to start
-//   - HYDRATING: Operation in progress, target node is pulling data
-//   - DEHYDRATING: The only state handled by source node, for cleanup after successful replication
-//   - **all other states**: Not reprocessed, require a new operation
-//
-// Returns only operations that should be actively processed by this node.
-func (p *FSMOpProducer) allOpsForNode(nodeId string) []ShardReplicationOp {
-	allNodeOps := p.fsm.GetOpsForNode(nodeId)
-
-	nodeOpsSubset := make([]ShardReplicationOp, 0, len(allNodeOps))
-	for _, op := range allNodeOps {
-		opState := p.fsm.GetOpState(op)
-
-		if opState.ShouldRestartOp() {
-			nodeOpsSubset = append(nodeOpsSubset, ShardReplicationOp{
-				ID: op.ID,
-				sourceShard: shardFQDN{
-					nodeId:       op.sourceShard.nodeId,
-					collectionId: op.sourceShard.collectionId,
-					shardId:      op.sourceShard.shardId,
-				},
-				targetShard: shardFQDN{
-					nodeId:       op.targetShard.nodeId,
-					collectionId: op.targetShard.collectionId,
-					shardId:      op.targetShard.shardId,
-				},
-			})
-		}
-	}
-
-	return nodeOpsSubset
-}
-
-// TimeProvider abstracts time operations to enable testing without time dependencies.
-type TimeProvider interface {
-	Now() time.Time
-}
-
-// RealTimeProvider implements the TimeProvider interface using the standard time package
-type RealTimeProvider struct{}
-
-// Now returns the current time
-func (p RealTimeProvider) Now() time.Time {
-	return time.Now()
-}
-
-// Timer defines an interface for scheduling tasks with a delay.
-type Timer interface {
-	// AfterFunc waits for the specified duration to elapse and then calls the provided function
-	// in its own goroutine.
-	AfterFunc(duration time.Duration, fn func()) *time.Timer
-}
-
-// CopyOpConsumer is an implementation of the OpConsumer interface that processes replication operations
-// by executing copy operations from a source shard to a target shard. It uses a ReplicaCopier to actually
-// carry out the copy operation. Moreover, it supports configurable backoff, timeout and concurrency limits.
-type CopyOpConsumer struct {
-	// logger is used for structured logging throughout the consumer's lifecycle.
-	// It provides detailed logs for each replication operation and any errors encountered.
-	logger *logrus.Entry
-
-	// leaderClient is responsible for interacting with the FSM to update the state of replication operations.
-	// It is used to update the status of operations during the replication process (e.g. update to HYDRATING state).
-	leaderClient types.FSMUpdater
-
-	// replicaCopier is used to handle the actual copying of replica data from the source shard to the target shard.
-	// It abstracts the mechanics of data replication and file copying.
-	replicaCopier types.ReplicaCopier
-
-	// backoffPolicy defines the retry mechanism for failed operations.
-	// It allows the consumer to retry replication operations using a backoff strategy in case of failure.
-	backoffPolicy backoff.BackOff
-
-	// maxWorkers sets the maximum number of concurrent workers that will be used to process replication operations.
-	// It controls the level of parallelism in the replication process allowing multiple replication operations to
-	// run concurrently.
-	maxWorkers int
-
-	// opTimeout defines the timeout duration for each replication operation.
-	// It ensures that operations do not hang indefinitely and are retried or terminated after the timeout period.
-	opTimeout time.Duration
-
-	// timeProvider abstracts time operations, allowing for easier testing and mocking of time-related functions.
-	timeProvider TimeProvider
-
-	// tokens controls the maximum number of concurrently running consumers
-	tokens chan struct{}
-
-	// nodeId uniquely identifies the node on which this consumer instance is running.
-	nodeId string
-}
-
-// String returns a string representation of the CopyOpConsumer,
-// including the node ID that uniquely identifies the consumer.
-//
-// The assumption is that each node runs one and only one replication engine,
-// which means there is one consumer per node.
-func (c *CopyOpConsumer) String() string {
-	return fmt.Sprintf("replication engine copy consumer on node '%s'", c.nodeId)
-}
-
-// NewCopyOpConsumer creates a new CopyOpConsumer instance responsible for executing
-// replication operations using a configurable worker pool.
-//
-// It uses a ReplicaCopier to perform the actual data copy.
-func NewCopyOpConsumer(
-	logger *logrus.Entry,
-	leaderClient types.FSMUpdater,
-	replicaCopier types.ReplicaCopier,
-	timeProvider TimeProvider,
-	nodeId string,
-	backoffPolicy backoff.BackOff,
-	opTimeout time.Duration,
-	maxWorkers int,
-) *CopyOpConsumer {
-	c := &CopyOpConsumer{
-		logger:        logger,
-		leaderClient:  leaderClient,
-		replicaCopier: replicaCopier,
-		backoffPolicy: backoffPolicy,
-		opTimeout:     opTimeout,
-		maxWorkers:    maxWorkers,
-		nodeId:        nodeId,
-		timeProvider:  timeProvider,
-		tokens:        make(chan struct{}, maxWorkers),
-	}
-	return c
-}
-
-// Consume processes replication operations from the input channel, ensuring that only a limited number of consumers
-// are active concurrently based on the maxWorkers value.
-func (c *CopyOpConsumer) Consume(ctx context.Context, in <-chan ShardReplicationOp) error {
-	c.logger.WithFields(logrus.Fields{
-		"node":    c.nodeId,
-		"workers": c.maxWorkers,
-	}).Info("starting replication operation consumer")
-
-	workerCtx, cancel := context.WithCancel(ctx)
-	defer cancel()
-
-	var wg sync.WaitGroup
-
-	for {
-		select {
-		case <-ctx.Done():
-			c.logger.WithField("reason", ctx.Err()).Info("context canceled, shutting down consumer")
-			wg.Wait() // Waiting for pending operations before terminating
-			return ctx.Err()
-
-		case op, ok := <-in:
-			if !ok {
-				c.logger.Info("operation channel closed, shutting down consumer")
-				wg.Wait() // Waiting for pending operations before terminating
-				return nil
-			}
-
-			select {
-			// The 'tokens' channel limits the number of concurrent workers (`maxWorkers`).
-			// Each worker acquires a token before processing an operation. If no tokens are available,
-			// the worker blocks until one is released. After completing the task, the worker releases the token,
-			// allowing another worker to proceed. This ensures only a limited number of workers is concurrently
-			// running replication operations and avoids overloading the system.
-			case c.tokens <- struct{}{}:
-
-				wg.Add(1)
-
-				// Here we capture the op argument used by the func below as the enterrors.GoWrapper requires calling
-				// a function without arguments.
-				operation := op
-
-				enterrors.GoWrapper(func() {
-					defer func() {
-						<-c.tokens // Release token when completed
-						wg.Done()
-					}()
-
-					opLogger := c.logger.WithFields(logrus.Fields{
-						"op":                operation.ID,
-						"source_node":       operation.sourceShard.nodeId,
-						"target_node":       operation.targetShard.nodeId,
-						"source_shard":      operation.sourceShard.shardId,
-						"target_shard":      operation.targetShard.shardId,
-						"source_collection": operation.sourceShard.collectionId,
-						"target_collection": operation.targetShard.collectionId,
-					})
-
-					opLogger.Info("worker processing replication operation")
-
-					// Start a replication operation with a timeout for completion to prevent replication operations
-					// from running indefinitely
-					opCtx, opCancel := context.WithTimeout(workerCtx, c.opTimeout)
-					defer opCancel()
-
-					err := c.processReplicationOp(opCtx, operation.ID, operation)
-					if err != nil && errors.Is(err, context.DeadlineExceeded) {
-						opLogger.WithError(err).Error("replication operation timed out")
-					} else if err != nil {
-						opLogger.WithError(err).Error("replication operation failed")
-					}
-				}, c.logger)
-
-			case <-ctx.Done():
-				return ctx.Err()
-			}
-		}
-	}
-}
-
-// processReplicationOp performs the full replication flow for a single operation.
-//
-// It performs of the following steps:
-//  1. Updates the operation status to HYDRATING using the leader FSM updater.
-//  2. Initiates the copy of replica data from the source node to the target shard.
-//  3. Once the copy succeeds, updates the sharding state to reflect the added replica.
-//
-// If any step fails, the operation is retried using the configured backoff policy.
-// Errors are logged and wrapped using the structured error group wrapper.
-func (c *CopyOpConsumer) processReplicationOp(ctx context.Context, workerId uint64, op ShardReplicationOp) error {
-	logger := c.logger.WithFields(logrus.Fields{
-		"op":                op.ID,
-		"source_node":       op.sourceShard.nodeId,
-		"target_node":       op.targetShard.nodeId,
-		"source_shard":      op.sourceShard.shardId,
-		"target_shard":      op.targetShard.shardId,
-		"source_collection": op.sourceShard.collectionId,
-		"target_collection": op.targetShard.collectionId,
-	})
-
-	startTime := c.timeProvider.Now()
-
-	eg := enterrors.NewErrorGroupWrapper(logger)
-	eg.Go(func() error {
-		return backoff.Retry(func() error {
-			if ctx.Err() != nil {
-				logger.WithError(ctx.Err()).Error("error while processing replication operation, shutting down")
-				return backoff.Permanent(ctx.Err())
-			}
-
-			if err := c.leaderClient.ReplicationUpdateReplicaOpStatus(op.ID, api.HYDRATING); err != nil {
-				logger.WithError(err).Error("failed to update replica status to 'HYDRATING'")
-				return err
-			}
-
-			logger.Info("starting replication copy operation")
-
-			if err := c.replicaCopier.CopyReplica(ctx, op.sourceShard.nodeId, op.sourceShard.collectionId, op.targetShard.shardId); err != nil {
-				logger.WithError(err).Error("failure while copying replica shard")
-				return err
-			}
-
-			if _, err := c.leaderClient.AddReplicaToShard(ctx, op.targetShard.collectionId, op.targetShard.shardId, op.targetShard.nodeId); err != nil {
-				logger.WithError(err).Error("failure while updating sharding state")
-				return err
-			}
-
-			c.logCompletedReplicationOp(workerId, startTime, c.timeProvider.Now(), op)
-
-			return nil
-		}, c.backoffPolicy)
-	})
-
-	return eg.Wait()
-}
-
-func (c *CopyOpConsumer) logCompletedReplicationOp(workerId uint64, startTime time.Time, endTime time.Time, op ShardReplicationOp) {
-	duration := endTime.Sub(startTime)
-
-	c.logger.WithFields(logrus.Fields{
-		"worker":            workerId,
-		"op":                op.ID,
-		"duration":          duration.String(),
-		"start_time":        startTime.Format(time.RFC1123),
-		"completed_since":   c.timeProvider.Now().Sub(endTime),
-		"source_node":       op.sourceShard.nodeId,
-		"target_node":       op.targetShard.nodeId,
-		"source_shard":      op.sourceShard.shardId,
-		"target_shard":      op.targetShard.shardId,
-		"source_collection": op.sourceShard.collectionId,
-		"target_collection": op.targetShard.collectionId,
-	}).Info("Replication operation completed successfully")
 }
